+++ src/app/comic/comic-list/comic-list.component.ts
@@ -40,13 +40,15 @@ export class ComicListComponent implements OnInit {
     }
 
     delete(comic: Comic) {
-        this.auth.openChallengePrompt(comic.title, `delete the comic "${comic.title}"`, () => {
+        this.auth.openChallengePrompt(comic.title, `delete the comic "${comic.title}"`).then((passed) => {
+            if (!passed) return;
             this.comicService.delete(comic);
         });
     }
 
     banOwner(comic: Comic) {
-        this.auth.openChallengePrompt(comic.title, `ban the owner of the comic "${comic.title}`, () => {
+        this.auth.openChallengePrompt(comic.title, `ban the owner of the comic "${comic.title}`).then((passed) => {
+            if (!passed) return;
             this.auth.ban(comic.accountID);
         });
     }
